Apps/requirements.txt
streamlit==1.24.1
python-dotenv==1.0.0
google-generativeai==0.2.1
Pillow==9.5.0
torch==2.0.1
transformers==4.30.2
diffusers==0.17.0
pyflakes==3.0.1